Fixes in type.toString
